+++ runtime/tests/component_adapter/test_execution_with_component_adapter_wiring.py
@@ -1,4 +1,5 @@
 import json
+import os
 from unittest import mock
 from uuid import UUID
 
@@ -9,6 +10,9 @@ from hetdesrun.backend.execution import (
     execute_transformation_revision,
 )
 from hetdesrun.models.execution import ExecByIdInput
+from hetdesrun.models.wiring import WorkflowWiring
+from hetdesrun.runtime.context import RuntimeExecutionContext
+from hetdesrun.trafoutils.trafo_collection import TrafoCollection
 
 
 @pytest.mark.asyncio
@@ -92,6 +96,57 @@ async def test_component_source_wiring_executes_via_backend_webservice(async_tes
     assert resp.status_code == 200
     assert resp.json()["error"] is None
     assert isinstance(resp.json()["output_results_by_output_name"]["output"], dict)
+
+
+@pytest.mark.asyncio
+@pytest.mark.usefixtures(
+    "_components_for_component_adapter_tests", "_pass_through_multits_component_in_db"
+)
+async def test_component_source_wiring_with_attrs_metadata_in_input_wiring(async_test_client):
+    """Tests that attrs provied with the input wiring are attached after loading from a true adapter
+
+    This must be tested with a real adapter an not direct_provisioning since the later does not
+    really parse the filter "value" itself. Instead it just returns the string value and parsing is
+    done later during workflow execution when the loaded obj is put into an input.
+    """
+    extra_metadata = {"some_metadata": {"a": 42, "b": "something"}}
+
+    exec_input = ExecByIdInput(
+        id=UUID("78ee6b00-9239-4214-b9bf-a093647f33f5"),  # pass through multits
+        wiring={
+            "input_wirings": [
+                {
+                    "workflow_input_name": "input",
+                    "adapter_id": "component-adapter",
+                    "ref_id": "f2a39f6b-3336-44f2-8c4f-2fd0a4651dd0",
+                    "filters": {
+                        "timestampFrom": "2025-01-14-12:00:00+00:00",
+                        "timestampTo": "2025-01-15-12:00:00+00:00",
+                        "frequency": "3h",
+                        "metrics": "a,b",
+                        "random_seed": 42,
+                        "metrics_parameters": r"{}",
+                    },
+                    "attrs": extra_metadata,
+                }
+            ]
+        },
+    )
+    with mock.patch(
+        "hetdesrun.adapters.component_adapter.config.component_adapter_config.allow_draft_components",
+        True,
+    ):
+        async with async_test_client as ac:
+            resp = await ac.post(
+                "/api/transformations/execute", json=json.loads(exec_input.model_dump_json())
+            )
+
+    assert resp.status_code == 200
+    assert resp.json()["error"] is None
+
+    result_obj_dict = resp.json()["output_results_by_output_name"]["output"]
+    assert isinstance(result_obj_dict, dict)
+    assert result_obj_dict["__metadata__"] == extra_metadata
 
 
 @pytest.mark.asyncio
@@ -211,3 +266,67 @@ async def test_component_sink_wiring_plotly_executes_correctly(tmpdir):
 
     assert "<html" in content
     assert "2020-01-01T01:15:27+00:00" in content
+
+
+@pytest.mark.asyncio
+async def test_runtime_exec_context_available_during_component_adapter_execution(
+    mocked_clean_test_db_session, async_test_client
+):
+    """
+    If component adapter is used, we want context information to be forwarded
+    into the execution of component adapter components.
+    """
+    with TrafoCollection(save_to_db=True) as tc:
+        pass_trough = tc.add_from_json_file(
+            os.path.join(
+                "transformations",
+                "components",
+                "connectors",
+                "pass-through_100_1946d5f8-44a8-724c-176f-16f3e49963af.json",
+            )
+        )
+        get_context_infos = tc.add_from_py_file(
+            os.path.join(
+                "tests",
+                "data",
+                "components",
+                "context_information.py",
+            )
+        )
+
+    async with async_test_client as ac:
+        exec_input = ExecByIdInput(
+            id=pass_trough.id,
+            job_id="bbbbbbbb-3cdf-45a4-98ad-bbbbbbbbbbbb",
+            wiring=WorkflowWiring(
+                input_wirings=[
+                    {
+                        "uri": f"hd://component-adapter/{get_context_infos.id}",
+                        "workflow_input_name": "input",
+                    }
+                ]
+            ),
+            runtime_execution_context=RuntimeExecutionContext(
+                hierarchy_object={
+                    "type": "blah",
+                    "id": "abc",
+                    "node_id": "def",
+                    "parent_node_id": "ghi",
+                }
+            ),
+        )
+        resp = await ac.post(
+            "/api/transformations/execute", json=json.loads(exec_input.model_dump_json())
+        )
+        assert resp.status_code == 200
+
+        resp_json = resp.json()
+
+        output = resp_json["output_results_by_output_name"]["output"]
+
+        assert output["hierarchy_object_info"] == {
+            "type": "blah",
+            "id": "abc",
+            "node_id": "def",
+            "parent_node_id": "ghi",
+        }