+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+## 0.13.9
+* new metadata convention and helper functions to access structured metadata for MULTITSFRAME and SERIES object handling fallback behaviour. See [docs](./docs/metadata_attrs.md) for details.
+* add support for frames / animations for rendering Plotly plots
+* add hetida platform data source base components
+
+## 0.13.8
+* Add hotkey for test execution dialog (Shift+Enter)
+* Add hotkey for quick test execution (using the current test wiring, without opening the execution/wiring dialog) (Alt+Enter)
+* Add hotkey (ESC) for closing the execution result / protocol view.
+* Add autosave status indicator
+* option to deprecate other revisions when releasing a new one
+* recommend semver patch version increase for version tag upon creating a new revision in frontend if previous revision has a simple semver version tag
+* add logging import + setup to default component code template
+
+## 0.13.7
+* compatibility fix for psycopg3 for structure upserting for large number of sources / sinks
+
+## 0.13.6
+* forward context into component adapter component execution
+
+## 0.13.5
+* Execution dialog now allows to specify timeranges with relative dtexp expressions.
+* New generic forecasting base component
+* Improved error handling for runtime service worker process termination
+* Add endpoint to obtain complete structure from backend structure service
+
+## 0.13.4
+* make metadata column adding more flexible and adaptable to different metadata structures
+
+## 0.13.3
+* add function in hdutils to add columns to a multitsframe from its metadata.
+* Virtual Structure adapter provides metadata
+* Input wirings support an attrs field that allows to update / override .attrs metadata for objects loaded from adapters
+* Execution/wiring dialog allows to enter time ranges directly, allowing possibly relative datetime expressions like `now-15min, now` (to e.g. specify interval of last 15 minutes) and more (see [dtexp](https://github.com/stewit/dtexp) docs)
+* Fixing validation in UI to allow entering `null` as manual input for optional inputs.
+* free text filters now can have default values.
+* Plotly config can now be specified under key `"config"` on the plotly fig dict object. The designer frontend will respect and use these configurations. Additionally, German Plotly locale package is installed into the frontend.
+
 ## 0.13.2
 * More lenient runtime execution context parsing